requirements.txt
Click==7.0
cmake==3.13.3
cycler==0.10.0
decorator==4.4.0
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
gunicorn==19.9.0
imageio==2.6.1
imutils==0.5.3
itsdangerous==1.1.0
Jinja2==2.10.3
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
networkx==2.4
numpy==1.17.3
opencv-python==4.1.1.26
Pillow==6.2.1
pyparsing==2.4.2
python-dateutil==2.8.0
PyWavelets==1.1.1
scikit-image==0.16.2
scipy==1.3.1
six==1.12.0
SQLAlchemy==1.3.10
Werkzeug==0.16.0
